# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberShield AI - Cybersecurity Awareness Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{userId}/progress:
    get:
      operationId: getUserProgress
      tags: [users]
      summary: Get a user's progress (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Progress"

  # COURSES
...
  /quizzes/course/{courseId}:
    get:
      operationId: listQuizzesByCourse
      tags: [quizzes]
      summary: Get all quiz questions for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
  /quizzes:
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a quiz question (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{quizId}:
    put:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update a quiz question (admin only)
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizUpdate"
      responses:
        "200":
          description: Updated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete a quiz question (admin only)
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /quizzes/submit/{courseId}:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      summary: Submit quiz answers for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"

  # PROGRESS
...
  /threats:
    get:
      operationId: listThreats
      tags: [threats]
      summary: List threats (own for user, all for admin)
      responses:
        "200":
          description: Threat reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreatReport"
    post:
      operationId: createThreat
      tags: [threats]
      summary: Submit a new threat report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreatInput"
      responses:
        "201":
          description: Created threat with AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatReport"
  /threats/{threatId}:
    get:
      operationId: getThreat
      tags: [threats]
      summary: Get a threat report by id
      parameters:
        - name: threatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Threat report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatReport"
  /threats/{threatId}/status:
    patch:
      operationId: updateThreatStatus
      tags: [threats]
      summary: Update threat report status (admin only)
      parameters:
        - name: threatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreatStatusUpdate"
      responses:
        "200":
          description: Updated threat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatReport"

  # ANNOUNCEMENTS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # User
...
    Course:
      type: object
      required: [id, title, description, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        content:
          type: string
        videoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Quiz:
      type: object
      required: [id, courseId, question, optionA, optionB, optionC, optionD, correctAnswer, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D]
        createdAt:
          type: string
          format: date-time
    QuizInput:
      type: object
      required: [courseId, question, optionA, optionB, optionC, optionD, correctAnswer]
      properties:
        courseId:
          type: integer
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D]
    QuizUpdate:
      type: object
      properties:
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
          enum: [A, B, C, D]
    QuizSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [quizId, answer]
            properties:
              quizId:
                type: integer
              answer:
                type: string
                enum: [A, B, C, D]
    QuizResult:
      type: object
      required: [score, total, passed, progress]
      properties:
        score:
          type: integer
        total:
          type: integer
        passed:
          type: boolean
        progress:
          $ref: "#/components/schemas/Progress"

    # Progress
    Progress:
      type: object
      required: [id, userId, courseId, score, completed, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        score:
          type: ["integer", "null"]
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        course:
          $ref: "#/components/schemas/Course"

    # Threat
    ThreatReport:
      type: object
      required: [id, userId, title, description, aiAnalysis, riskLevel, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: string
        link:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        aiAnalysis:
          type: string
        riskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        status:
          type: string
          enum: [PENDING, REVIEWED, RESOLVED]
        createdAt:
          type: string
          format: date-time
...
    ThreatInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
          minLength: 10
        link:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    ThreatStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [PENDING, REVIEWED, RESOLVED]

    # Announcement
...