# lib/api-zod/src/generated/types/quizInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberShield AI - Cybersecurity Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizInputCorrectAnswer } from './quizInputCorrectAnswer';

export interface QuizInput {
  courseId: number;
  question: string;
  optionA: string;
  optionB: string;
  optionC: string;
  optionD: string;
  correctAnswer: QuizInputCorrectAnswer;
}