# lib/api-zod/src/generated/types/threatStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberShield AI - Cybersecurity Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatStatusUpdateStatus } from './threatStatusUpdateStatus';

export interface ThreatStatusUpdate {
  status: ThreatStatusUpdateStatus;
}